Create google_cloud_speech.py
I tried to make a script to use Google Cloud Speech to transcribe text. 
Google cloud speech requires credentials. 
I created a service account on the Google Cloud Platform, and downloaded a key in the form of a JSON file.
I don't know how to point to that key file in my script, line 14